[Refactor] Extract network schemas
